fix: entity class pk long으로 수정 및 quizChoice 연관관계 수정

+++ BackEnd/chartflow/src/main/java/com/ssafy/chartflow/board/entity/Article.java
@@ -19,7 +19,7 @@ public class Article {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "article_id")
-    private int articleId;
+    private long articleId;
 
     @ManyToOne
     @JoinColumn(name = "user_id")
